fixed lexing error which complained about the features file
There was nothing wrong with the cucumber file format. The issue was that I copied the steps over from the Ruby framework which had a different file encoding (Should be UTF-8). since the zero width space character ' \u200b' was present the lexing error was triggered as it thinks there's a syntactical error, showing confusion as to the end of the feature file.

Whenever this happens
1) cut and paste the features file in IntelliJ onto notepad
2) highlight text, right click 'show Unicode control characters'
3) select all and copy
4) paste back into the feature file in IntelliJ
5) save and re-run the code
6) by the way it is good practice to show line numbers on the file as it might reveal empty new lines which can trigger this
7) always make sure that when you CTRL + click on the steps on .feature that they auto bring up the step defs